refactor: 플랜 도메인 엔티티 클래스 BaseTimeEntity 변경 (#77)
