+++ src/main/kotlin/kr/wooco/woocobe/plan/infrastructure/storage/entity/PlanCategoryJpaEntity.kt
@@ -1,11 +1,11 @@
 package kr.wooco.woocobe.plan.infrastructure.storage.entity
 
-import io.hypersistence.utils.hibernate.id.Tsid
 import jakarta.persistence.Column
 import jakarta.persistence.Entity
 import jakarta.persistence.Id
 import jakarta.persistence.Table
-import kr.wooco.woocobe.common.infrastructure.storage.BaseTimeEntity
+import kr.wooco.woocobe.common.infrastructure.storage.Tsid
+import kr.wooco.woocobe.common.infrastructure.storage.entity.BaseTimeEntity
 
 @Entity
 @Table(name = "plan_categories")
@@ -16,5 +16,5 @@ class PlanCategoryJpaEntity(
     val name: String,
     @Id @Tsid
     @Column(name = "plan_category_id")
-    val id: Long? = 0L,
+    override val id: Long = 0L,
 ) : BaseTimeEntity()
