+++ src/main/kotlin/kr/wooco/woocobe/plan/infrastructure/storage/entity/PlanPlaceJpaEntity.kt
@@ -1,11 +1,11 @@
 package kr.wooco.woocobe.plan.infrastructure.storage.entity
 
-import io.hypersistence.utils.hibernate.id.Tsid
 import jakarta.persistence.Column
 import jakarta.persistence.Entity
 import jakarta.persistence.Id
 import jakarta.persistence.Table
-import kr.wooco.woocobe.common.infrastructure.storage.BaseTimeEntity
+import kr.wooco.woocobe.common.infrastructure.storage.Tsid
+import kr.wooco.woocobe.common.infrastructure.storage.entity.BaseTimeEntity
 
 @Entity
 @Table(name = "plan_places")
@@ -17,5 +17,5 @@ class PlanPlaceJpaEntity(
     @Column(name = "place_id")
     val placeId: Long,
     @Id @Tsid
-    val id: Long? = 0L,
+    override val id: Long = 0L,
 ) : BaseTimeEntity()
